+++ pyidi/pyidi.py
@@ -11,7 +11,6 @@ warnings.simplefilter("default")
 from .methods import IDIMethod, SimplifiedOpticalFlow, LucasKanade, DirectionalLucasKanade #, LucasKanadeSc, LucasKanadeSc2, GradientBasedOpticalFlow
 from .video_reader import VideoReader
 from . import tools
-from .GUIs import selection
 
 available_method_shortcuts = [
     ('sof', SimplifiedOpticalFlow),